requirements.txt
python-telegram-bot==11.1
redis==4.1.4
vk-api==11.9.7
python-dotenv==0.19.2